Rooba | Refactors grok_exporter and pulls out config
